+++ lib/oli_web/controllers/page_delivery_controller.ex
@@ -6,7 +6,6 @@ defmodule OliWeb.PageDeliveryController do
       is_section_instructor_or_admin?: 2
     ]
 
-  alias Oli.Delivery.Student.Summary
   alias Oli.Delivery.Page.PageContext
   alias Oli.Delivery.Sections
   alias Oli.Rendering.Context
@@ -20,17 +19,64 @@ defmodule OliWeb.PageDeliveryController do
   alias Oli.Resources.ResourceType
   alias Oli.Grading
   alias Oli.PartComponents
+  alias Oli.Rendering.Activity.ActivitySummary
+
+  def index_preview(conn, %{"section_slug" => section_slug}) do
+    user = conn.assigns.current_user
+    current_author = conn.assigns.current_author
+    is_admin? = !is_nil(current_author) and Oli.Accounts.is_admin?(current_author)
+
+    # We only allow access to preview mode if the user is logged in as an author admin, or
+    # is an instructor enrolled in the course section.  If the user is enrolled as a student,
+    # we want to redirect out of this mode to render the page in regular delivery mode.
+
+    if is_admin? or Sections.is_enrolled?(user.id, section_slug) do
+      if is_admin? or Sections.has_instructor_role?(user, section_slug) do
+        case Sections.get_section_by(slug: section_slug)
+             |> Oli.Repo.preload([:base_project, :root_section_resource]) do
+          nil ->
+            render(conn, "error.html")
+
+          section ->
+            hierarchy = Oli.Publishing.DeliveryResolver.full_hierarchy(section.slug)
+
+            render(conn, "index.html",
+              title: section.title,
+              description: section.description,
+              section_slug: section_slug,
+              hierarchy: hierarchy,
+              preview_mode: true
+            )
+        end
+      else
+        # Any attempt by a student to enter preview mode is simply ignored and redirect back to
+        # regular delivery mode
+        redirect(conn, to: Routes.page_delivery_path(conn, :index, section_slug))
+      end
+    else
+      render(conn, "not_authorized.html")
+    end
+  end
 
   def index(conn, %{"section_slug" => section_slug}) do
     user = conn.assigns.current_user
 
     if Sections.is_enrolled?(user.id, section_slug) do
-      case Summary.get_summary(section_slug, user) do
-        {:ok, summary} ->
-          render(conn, "index.html", section_slug: section_slug, summary: summary)
-
-        {:error, _} ->
+      case Sections.get_section_by(slug: section_slug)
+           |> Oli.Repo.preload([:base_project, :root_section_resource]) do
+        nil ->
           render(conn, "error.html")
+
+        section ->
+          hierarchy = Oli.Publishing.DeliveryResolver.full_hierarchy(section.slug)
+
+          render(conn, "index.html",
+            title: section.title,
+            description: section.description,
+            section_slug: section_slug,
+            hierarchy: hierarchy,
+            preview_mode: false
+          )
       end
     else
       render(conn, "not_authorized.html")
@@ -51,15 +97,123 @@ defmodule OliWeb.PageDeliveryController do
     end
   end
 
+  def page_preview(conn, %{
+        "section_slug" => section_slug,
+        "revision_slug" => revision_slug
+      }) do
+    user = conn.assigns.current_user
+    current_author = conn.assigns.current_author
+    is_admin? = !is_nil(current_author) and Oli.Accounts.is_admin?(current_author)
+
+    # We only allow access to preview mode if the user is logged in as an author admin, or
+    # is an instructor enrolled in the course section.  If the user is enrolled as a student,
+    # we want to redirect out of this mode to render the page in regular delivery mode.
+
+    if is_admin? or Sections.is_enrolled?(user.id, section_slug) do
+      if is_admin? or Sections.has_instructor_role?(user, section_slug) do
+        revision = Oli.Publishing.DeliveryResolver.from_revision_slug(section_slug, revision_slug)
+        render_preview_mode(conn, section_slug, revision)
+      else
+        # Any attempt by a student to enter preview mode is simply ignored and redirect back to
+        # regular delivery mode
+        redirect(conn, to: Routes.page_delivery_path(conn, :page, section_slug, revision_slug))
+      end
+    else
+      render(conn, "not_authorized.html")
+    end
+  end
+
   def page(conn, %{"section_slug" => section_slug, "revision_slug" => revision_slug}) do
     user = conn.assigns.current_user
 
     if Sections.is_enrolled?(user.id, section_slug) do
       PageContext.create_for_visit(section_slug, revision_slug, user)
-      |> render_page(conn, section_slug, user)
+      |> render_page(conn, section_slug, user, false)
     else
       render(conn, "not_authorized.html")
     end
+  end
+
+  defp render_preview_mode(
+         conn,
+         section_slug,
+         %{content: %{"advancedDelivery" => true}} = revision
+       ) do
+    user = conn.assigns.current_user
+
+    if Sections.is_instructor?(user, section_slug) do
+      PageContext.create_for_visit(section_slug, revision.slug, user)
+      |> render_page(conn, section_slug, user, true)
+    else
+      render(conn, "not_authorized.html")
+    end
+  end
+
+  defp render_preview_mode(conn, section_slug, revision) do
+    section = conn.assigns.section
+
+    page_model = Map.get(revision.content, "model")
+
+    type_by_id =
+      Activities.list_activity_registrations()
+      |> Enum.reduce(%{}, fn e, m -> Map.put(m, e.id, e) end)
+
+    activity_ids =
+      Oli.Resources.PageContent.flat_filter(revision.content, fn item ->
+        item["type"] == "activity-reference"
+      end)
+      |> Enum.map(fn %{"activity_id" => id} -> id end)
+
+    hierarchy = Oli.Publishing.DeliveryResolver.full_hierarchy(section_slug)
+    {previous, next} = Oli.Delivery.Page.PageContext.determine_previous_next(hierarchy, revision)
+
+    activity_map =
+      Oli.Publishing.DeliveryResolver.from_resource_id(section_slug, activity_ids)
+      |> Enum.map(fn rev ->
+        type = Map.get(type_by_id, rev.activity_type_id)
+
+        %ActivitySummary{
+          id: rev.resource_id,
+          script: type.authoring_script,
+          attempt_guid: nil,
+          state: nil,
+          model: Jason.encode!(rev.content) |> Oli.Delivery.Page.ActivityContext.encode(),
+          delivery_element: type.delivery_element,
+          authoring_element: type.authoring_element,
+          graded: revision.graded
+        }
+      end)
+      |> Enum.reduce(%{}, fn r, m -> Map.put(m, r.id, r) end)
+
+    all_activities = Activities.list_activity_registrations()
+
+    render_context = %Context{
+      user: conn.assigns.current_user,
+      section_slug: section_slug,
+      mode: :instructor_preview,
+      activity_map: activity_map,
+      activity_types_map: Enum.reduce(all_activities, %{}, fn a, m -> Map.put(m, a.id, a) end)
+    }
+
+    html = Page.render(render_context, page_model, Page.Html)
+
+    conn = put_root_layout(conn, {OliWeb.LayoutView, "page.html"})
+
+    render(
+      conn,
+      "instructor_preview.html",
+      %{
+        summary: %{title: section.title},
+        section_slug: section_slug,
+        scripts: Enum.map(all_activities, fn a -> a.authoring_script end),
+        preview_mode: true,
+        previous_page: previous,
+        next_page: next,
+        title: revision.title,
+        html: html,
+        objectives: []
+      }
+    )
   end
 
   defp render_page(
@@ -71,6 +225,7 @@ defmodule OliWeb.PageDeliveryController do
          } = context,
          conn,
          section_slug,
+         _,
          _
        ) do
     # Only consider graded attempts
@@ -120,7 +275,8 @@ defmodule OliWeb.PageDeliveryController do
            context,
          conn,
          section_slug,
-         _
+         _,
+         preview_mode
        ) do
     layout =
       case Map.get(context.page.content, "displayApplicationChrome", true) do
@@ -157,20 +313,25 @@ defmodule OliWeb.PageDeliveryController do
       previous_page: context.previous_page,
       next_page: context.next_page,
       user_id: user.id,
-      preview_mode: false
+      preview_mode: preview_mode
     })
   end
 
-  defp render_page(%PageContext{progress_state: :error}, conn, _, _) do
+  defp render_page(%PageContext{progress_state: :error}, conn, _, _, _) do
     render(conn, "error.html")
   end
 
   # This case handles :in_progress and :revised progress states
-  defp render_page(%PageContext{} = context, conn, section_slug, user) do
+  defp render_page(%PageContext{} = context, conn, section_slug, user, _) do
     render_context = %Context{
       user: user,
       section_slug: section_slug,
-      review_mode: context.review_mode,
+      mode:
+        if context.review_mode do
+          :review
+        else
+          :delivery
+        end,
       activity_map: context.activities
     }
 
@@ -250,7 +411,7 @@ defmodule OliWeb.PageDeliveryController do
 
     if Sections.is_enrolled?(user.id, section_slug) do
       PageContext.create_for_review(section_slug, attempt_guid, user)
-      |> render_page(conn, section_slug, user)
+      |> render_page(conn, section_slug, user, false)
     else
       render(conn, "not_authorized.html")
     end